% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_name_chunks.R
\name{rmd_name_chunks}
\alias{rmd_name_chunks}
\title{Read an Rmd rmarkdown file and apply enumerated code chunks}
\usage{
rmd_name_chunks(fn_in = NULL, fn_out = NULL, prefix_chunk_name = "chunk-")
}
\arguments{
\item{fn_in}{Rmd filename in}

\item{fn_out}{Rmd filename out.  If NULL, then this is assigned \code{fn_in}.}
}
\value{
NULL, invisibly
}
\description{
Read an Rmd rmarkdown file and apply enumerated code chunks
}
\examples{
rmd_name_chunks(fn_in = "test_in.Rmd", fn_out = "test_out.Rmd", prefix_chunk_name = "chunk-")
}
