% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RmdNameChunk.R
\docType{package}
\name{RmdNameChunk}
\alias{RmdNameChunk}
\alias{RmdNameChunk-package}
\title{RmdNameChunk: Read an Rmd rmarkdown file and apply enumerated code chunks}
\description{
RmdNameChunk: Read an Rmd rmarkdown file and apply enumerated code chunks
}
